refactor: reorganize components and move roles and projects data to config
