+++ src/pages/index.astro
@@ -7,7 +7,7 @@ import {
   Projects,
   Blog,
   Categories,
-} from "../components/home";
+} from "../modules/home";
 
 import { BLOG_TITLE, BLOG_DESCRIPTION } from "../config";
 ---
@@ -28,7 +28,6 @@ import { BLOG_TITLE, BLOG_DESCRIPTION } from "../config";
     </section>
 
     <section>
-      {/* Blog & Categories Section */}
       <section class="flex flex-col md:flex-row gap-4">
         <div class="w-full md:w-1/2">
           <Blog />
